feat & refactor: 1. Merge adCreationRequest.go and adCondition.go. 2. Implement ad query function. 3. Refactor integers to use int32 or int64 to ensure type correctness. 4. Update gitignore to ignore txt files and test files.

+++ controller/Advertisements.go
@@ -4,8 +4,8 @@ import "github.com/lib/pq"
 
 type Advertisements struct {
 	Title     string         `json:"title"`
-	Start_At  int32          `json:"start_at" gorm:"index:idx_ad_start_end_age, priority:2"`
-	End_At    int32          `json:"end_at" gorm:"index:idx_ad_start_end_age, priority:1"`
+	Start_At  int64          `json:"start_at" gorm:"index:idx_ad_start_end_age, priority:2"`
+	End_At    int64          `json:"end_at" gorm:"index:idx_ad_start_end_age, priority:1"`
 	Age_Start int32          `json:"age_start,omitempty" gorm:"index"`
 	Age_End   int32          `json:"age_end,omitempty" gorm:"index"`
 	Gender    pq.StringArray `json:"gender,omitempty" gorm:"type:text[]"`
